+++ Gateway/Request/PaymentOffsiteBuilder.php.orig
@@ -4,9 +4,9 @@ namespace Omise\Payment\Gateway\Request;
 use Magento\Framework\UrlInterface;
 use Magento\Payment\Gateway\Helper\SubjectReader;
 use Magento\Payment\Gateway\Request\BuilderInterface;
-use Omise\Payment\Observer\OffsiteAlipayDataAssignObserver;
+use Omise\Payment\Observer\OffsiteInternetbankingDataAssignObserver;
 
-class PaymentAlipayBuilder implements BuilderInterface
+class PaymentOffsiteBuilder implements BuilderInterface
 {
     /**
      * @var string
@@ -39,8 +39,8 @@ class PaymentAlipayBuilder implements BuilderInterface
         $method  = $payment->getPayment();
 
         return [
-            self::OFFSITE    => 'alipay',
-            self::RETURN_URI => $this->url->getUrl('omise/callback/offsite', ['_secure' => true])
+            self::OFFSITE    => $method->getAdditionalInformation(OffsiteInternetbankingDataAssignObserver::OFFSITE),
+            self::RETURN_URI => $this->url->getUrl('omise/callback/internetbanking', ['_secure' => true])
         ];
     }
 }
